Huge Code Cleanup.
Since this is dead I might as well. Also fixes some possible crashes.

+++ src/main/java/me/trambled/ozark/ozarkclient/guiscreen/hud/pinnables/InventoryPreview.java
@@ -26,8 +26,8 @@ public class InventoryPreview extends Pinnable {
 			for (int i = 0; i < 27; i++) {
 				ItemStack item_stack = mc.player.inventory.mainInventory.get(i + 9);
 
-				int item_position_x = (int) this.get_x() + (i % 9) * 16;
-				int item_position_y = (int) this.get_y() + (i / 9) * 16;
+				int item_position_x = this.get_x() + (i % 9) * 16;
+				int item_position_y = this.get_y() + (i / 9) * 16;
 
 				mc.getRenderItem().renderItemAndEffectIntoGUI(item_stack, item_position_x, item_position_y);
 				mc.getRenderItem().renderItemOverlayIntoGUI(mc.fontRenderer, item_stack, item_position_x, item_position_y, null);
